Fix: new banner and description added
new images folder created
